Update Board.java
Hey, Nice code!  just a comment regarding line 96. The methods below that line are setters but they are actually getters. in the name. Nothing serious but for keeping good practices. 
So my proposal is just to change line 96 where it says setters to getters.
For future events, bear in mind that getters are always NOT void whereas setters will ALWAYS be void.